refactor(document generation): move generation into seperate class
ref #584
